Vagrant: add internal overlayfs for NPM's node_modules folder
Using shared folders can be a problem in VirtualBox Guest Addition 6.x.
This is discussed for instance in these places;

  https://github.com/laravel/homestead/issues/1239
  https://www.virtualbox.org/ticket/18776
  https://www.virtualbox.org/ticket/19086

This is bug likely bites us when we so far tried to have a Ubuntu 18.04
Vagrant VM. Since changing the version of the Guest Additions seems to
tricky in this context, an overlay filesystem allows to ship around the
problem by mapping the node_modules folder within the VM to a separate
internal folder. This allows us a) to not interfer with the host's
node_modules folder and b) to write data without a shared filesystem and
not running into this error.

This adds an fstab entry for the overlay so that this change survives
reboot.